pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evaluation-first-attention"
version = "0.1.0"
description = "Evaluation-First Generation: Specification-Driven LLM Output Quality via Dynamic Rubric Conditioning and Failure-Weighted Reattention"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Karthick Raja M", email = "karthickrajam18@gmail.com"},
]

dependencies = [
    "litellm>=1.50.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
]
experiments = [
    "pandas>=2.0",
    "numpy>=1.26",
    "tiktoken>=0.7",
    "python-dotenv>=1.0",
    "scipy>=1.12",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
